requirements.txt
asgiref
coverage
dj-database-url
Django
djangorestframework
gunicorn
packaging
psycopg2-binary
python-dotenv
sqlparse
whitenoise
pillow
django-widget-tweaks
